+++ packages/expo-cli/lib/test/configure-plugin.test.js
@@ -1,6 +1,7 @@
 const withFixture = require('./lib/with-fixture')
 const configurePlugin = require('../configure-plugin')
 const { readFile } = require('fs/promises')
+const { blue } = require('kleur')
 
 describe('expo-cli: upload sourcemaps configure-plugin', () => {
   it('should work on a fresh project', async () => {
@@ -10,25 +11,58 @@ describe('expo-cli: upload sourcemaps configure-plugin', () => {
 
       const appJsonRaw = await readFile(`${projectRoot}/app.json`, 'utf8')
       const appJson = JSON.parse(appJsonRaw)
-
       expect(appJson.expo.plugins).toContain('@bugsnag/plugin-expo-eas-sourcemaps')
+
+      const packageJsonRaw = await readFile(`${projectRoot}/package.json`, 'utf8')
+      const packageJson = JSON.parse(packageJsonRaw)
+      expect(packageJson.scripts['eas-build-on-success']).toContain('npx bugsnag-eas-build-on-success')
     })
   })
 
   it('shouldn’t duplicate the hook config', async () => {
     await withFixture('already-installed-02', async (projectRoot) => {
-      const msg = await configurePlugin(projectRoot)
-      expect(msg).toMatch(/ is already installed/)
+      const logSpy = jest.spyOn(console, 'log').mockImplementation(() => {})
+
+      await configurePlugin(projectRoot)
+      expect(logSpy).toHaveBeenCalledWith(blue('Plugin is already configured in app.json'))
 
       const appJsonRaw = await readFile(`${projectRoot}/app.json`, 'utf8')
       const appJson = JSON.parse(appJsonRaw)
-
       expect(appJson.expo.plugins.length).toBe(1)
     })
   })
 
-  it('should create a basic file when there is no app.json', async () => {
+  it('shouldn’t duplicate the EAS build hook', async () => {
+    await withFixture('already-installed-01', async (projectRoot) => {
+      const logSpy = jest.spyOn(console, 'log').mockImplementation(() => {})
+
+      await configurePlugin(projectRoot)
+      expect(logSpy).toHaveBeenCalledWith(blue('EAS Build hook already configured in package.json'))
+
+      const packageJsonRaw = await readFile(`${projectRoot}/package.json`, 'utf8')
+      const packageJson = JSON.parse(packageJsonRaw)
+      expect(packageJson.scripts['eas-build-on-success']).toStrictEqual('npx bugsnag-eas-build-on-success')
+    })
+  })
+
+  it('should chain to an existing EAS build hook if present', async () => {
+    await withFixture('already-installed-02', async (projectRoot) => {
+      await configurePlugin(projectRoot)
+
+      const packageJsonRaw = await readFile(`${projectRoot}/package.json`, 'utf8')
+      const packageJson = JSON.parse(packageJsonRaw)
+      expect(packageJson.scripts['eas-build-on-success']).toStrictEqual('pre-existing-command && npx bugsnag-eas-build-on-success')
+    })
+  })
+
+  it('should provide a reasonable error when there is no package.json', async () => {
     await withFixture('empty-00', async (projectRoot) => {
+      await expect(configurePlugin(projectRoot)).rejects.toThrow(/Couldn’t find package\.json/)
+    })
+  })
+
+  it('should create a basic file when there is no app.json', async () => {
+    await withFixture('empty-01', async (projectRoot) => {
       const msg = await configurePlugin(projectRoot)
       expect(msg).toBe(undefined)